Cast length to uintmax_t before shifting.
On systems where size_t is 32 bits in size, a shift of 32 is already
undefined. By increasing the width to uintmax_t first, it should
succeed, as uintmax_t is supposed to be at least 64 bits on POSIX
systems.

+++ src/libc/unistd/crypt_hash_sha512.h
@@ -168,7 +168,7 @@ static void SHA512_Update(SHA512_CTX *ctx, const void *data, size_t len) {
   }
 
   // Update the length counter.
-  uint64_t length[2] = {len >> 32 >> 32, len};
+  uint64_t length[2] = {(uintmax_t)len >> 32 >> 32, len};
   ctx->length[0] += length[0];
   if (add_overflow(ctx->length[1], length[1], &ctx->length[1]))
     ++ctx->length[0];
